+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * :sparkles: Stringx orEmpty, orDefault, emptyToNull, blankToNull support CharSequence
 * :sparkles: Formatx add hiddenStartChars, hiddenMiddleChars, hiddenEndChars method
 * :sparkles: Zipx.kt zipCompressFilesTo method add Overload method
+* :sparkles: Premisex.kt requireNotNull and checkNotNull method add Overload method
 
 ## 0.6.5
 * :hammer: String orContains and andContains change to extension method